+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### [Mailer]
 
 - Added the soft delete of mail type variants. remp/crm#2721
+- Add ability to log bots use in Mailgun "opened" events via standalone Hermes handler (disabled by default). remp/analytika#137
 
 ## [1.1.0] - 2023-01-27